Co-locate cloud sync registry integration
